fix: mapping `Number` types to POJO query output

+++ client-core/src/test/java/com/influxdb/query/internal/FluxResultMapperTest.java
@@ -203,6 +203,6 @@ class FluxResultMapperTest {
         @Column
         private Double fieldDouble;
         @Column
-        private Double fieldInt;
+        private Integer fieldInt;
     }
 }
